requirements.txt
torch
torchvision
opencv-python
imutils
Pillow
numpy
